+++ hast_web/ws_server.cpp
@@ -143,14 +143,14 @@ WebSocketFrameType ws_server::getFrame(unsigned char* in_buffer, int in_length, 
 	return ERROR_FRAME;
 }
 
-inline void ws_server::echo_back_msg(const short int socket_index, std::string &msg){
+inline void ws_server::echo_back_msg(const int socket_index, std::string &msg){
 	int len {msg.length()+1};
 	char buf[len];
 	len = makeFrame(TEXT_FRAME,msg.c_str(),len-1,buf,len);
 	send(socket_index, buf, len,0);
 }
 
-inline void ws_server::echo_back_msg(const short int socket_index, const char* msg){
+inline void ws_server::echo_back_msg(const int socket_index, const char* msg){
 	int len {strlen(msg)+1};
 	char buf[len];
 	len = makeFrame(TEXT_FRAME, msg, len-1,buf,len);
